Moved Map class to the Library for a common access
