--- Assets/AdminMap/Scripts/GamePanel/Map.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4472b3e1f977b944e830a371734b2cf3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
